fix: fixed the test_result update

+++ src/infrastructure/db/entities/test_answer.entity.ts
@@ -21,7 +21,7 @@ export class TestAnswerEntity {
   @IsNotEmpty()
   isCorrect: boolean;
 
-  @Column({ type: 'int', default: 1 })
+  @Column({ type: 'float', default: 1 })
   @IsNotEmpty()
   score: number;
